+++ changelog.rst
@@ -1,7 +1,16 @@
 Changelog
 =========
 
-Last change: 06-JUN-2024 MTS
+Last change: 06-JUL-2024 MTS
+
+0.7.0
+-----
+- Adaptive Intersection Maximization (AIM, doi: 10.1038/s41592-022-01307-0) implemented
+- Z fitting improved by setting bounds on fitted z values to avoid NaNs
+- CMD ``clusterfile`` fixed 
+- Picasso: Render 3D, rectangular and polygonal pick fixed
+- picasso.localize.localize fixed
+- default MLE fitting uses different sx and sy (CMD only)
 
 0.6.9 - 0.6.11
 --------------